Added generator random

+++ EVONN/main.cpp
@@ -2,9 +2,9 @@
 
 int main() {
 	BrainSchema shema;
-	shema.addConstant(Data(0));
-	shema.addConstant(Data(1));
-	shema.addConstant(Data(-1));
+	shema.addConstant(Data((int)0));
+	shema.addConstant(Data((int)1));
+	shema.addConstant(Data((int)-1));
 	for (int i = 0; i < 10; ++i) {
 		shema.addInput(DT_INTEGER);
 	}
